+++ src/theme/DocSidebar/Desktop/index.tsx
@@ -25,14 +25,14 @@ function DocSidebarDesktop({ path, sidebar, onCollapse, isHidden }: Props) {
   } = useThemeConfig();
 
   const location = useLocation();
-  const isGatorDocs = location.pathname.startsWith('/gator');
+  const isGatorDocs = location.pathname.startsWith('/delegation-toolkit');
 
   let versionDropdown = null;
   try {
     if (isGatorDocs) {
       versionDropdown = (
         <div className={styles.versionDropdownContainer}>
-          <SidebarVersionDropdown pluginId="gator" />
+          <SidebarVersionDropdown path="delegation-toolkit" />
         </div>
       );
     }