+++ docs/whats-new.md
@@ -11,7 +11,7 @@ of the [MetaMask developer page](https://metamask.io/developer/).
 
 ## June 2025
 
-- Added [Delegation Toolkit](/gator) to MetaMask documentation.
+- Added [Delegation Toolkit](/delegation-toolkit) to MetaMask documentation.
   ([#2045](https://github.com/MetaMask/metamask-docs/pull/2045))
 
 ## May 2025
@@ -23,7 +23,7 @@ of the [MetaMask developer page](https://metamask.io/developer/).
   ([#1940](https://github.com/MetaMask/metamask-docs/pull/1940))
 - Documented [MetaMask SDK + Web3Auth SDK integration](/sdk/quickstart/javascript-web3auth).
   ([#2029](https://github.com/MetaMask/metamask-docs/pull/2029))
-- Added [Delegation Toolkit](/gator) to MetaMask documentation.
+- Added [Delegation Toolkit](/delegation-toolkit) to MetaMask documentation.
   ([#2045](https://github.com/MetaMask/metamask-docs/pull/2045))
 - Documented [how to use the Snaps sandbox](/snaps/how-to/test-a-snap/#test-in-the-sandbox).
   ([#2030](https://github.com/MetaMask/metamask-docs/pull/2030))